Updated contributors.
Changed all instances and references of $all_custom_fields to $all_custom_sets.
Now the wording is more descriptive.
Also, improved the way this array is populated after it comes out of the db.

Started working on different inputs for the search form.
This is more of a WIP, that functionality is broken in this commit.

Added a patch submitted by Steve Dickinson, that might have been my problem on glz_article
and glz_article_custom as mentioned by Dale a few weeks back.